requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
urllib3>=2.0.0
openai>=1.12.0
anthropic>=0.18.0
ollama>=0.1.6
aiohttp>=3.9.0
asyncio>=3.4.3
selenium>=4.15.0
webdriver-manager>=4.0.0
httpx>=0.26.0
dnspython>=2.4.0
python-whois>=0.8.0
shodan>=1.30.0
reportlab>=4.0.0
jinja2>=3.1.0
markdown>=3.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
colorama>=0.4.6
tabulate>=0.9.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
websocket-client>=1.6.0
cryptography>=41.0.0
python-jwt>=4.0.0
sqlalchemy>=2.0.0
sqlite3worker>=1.1.7
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0
validators>=0.22.0
python-magic>=0.4.27
loguru>=0.7.0
python-dateutil>=2.8.0
packaging>=23.0
google-generativeai>=0.4.0
rich>=13.7.0
